Commiting all local changes

+++ Paylocity.Business/EmployeeBL.cs
@@ -12,10 +12,12 @@ namespace Paylocity.Business
     public class EmployeeBL : IEmployeeBL
     {
         private readonly IEmployeeDAL _employeeDal;
+        private readonly IDependentDAL _dependentDal;
         private readonly IUnitOfWorkFactory _unitOfWorkFactory;
-        public EmployeeBL(IUnitOfWorkFactory unitOfWorkFactory, IEmployeeDAL employeeDal)
+        public EmployeeBL(IUnitOfWorkFactory unitOfWorkFactory, IEmployeeDAL employeeDal, IDependentDAL dependentDal)
         {
             _employeeDal = employeeDal;
+            _dependentDal = dependentDal;
             _unitOfWorkFactory = unitOfWorkFactory;
         }
         public async Task<EmployeeModel> GetEmployee(int id)
@@ -38,9 +40,14 @@ namespace Paylocity.Business
         {
             using (var unitOfWork = _unitOfWorkFactory.CreateUnitOfWork())
             {
-                var isDeleted = await _employeeDal.DeleteEmployee(unitOfWork, id);
+                // delete all the dependents associated with the employee first
+                var isDeleted = await _dependentDal.DeleteDependentsByEmployeeId(unitOfWork, id);
+                if (!isDeleted)
+                    return false;
+                await _employeeDal.DeleteEmployee(unitOfWork, id);
+                //complete the transaction
                 await unitOfWork.Complete();
-                return isDeleted;
+                return true;
             }
         }
